pyproject.toml
[project]
name = "trackers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "flask>=3.1.2",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
]

[project.scripts]
trackers-app = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "beautifulsoup4>=4.14.3",
    "hypothesis>=6.148.2",
    "pytest>=9.0.0",
]
